requirements.txt
Django==5.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.0
gunicorn
dj-database-url
whitenoise==6.6.0
pytz==2024.1
psycopg2-binary==2.9.11
fastapi
uvicorn[standard]
paho-mqtt